fix: defaults in configs

+++ src/config.rs
@@ -23,21 +23,22 @@ pub struct AppConfig {
 }
 
 #[derive(Clone, Debug, serde::Deserialize, serde::Serialize, SettingsNested, smart_default::SmartDefault)]
+#[serde(default)]
 pub struct Outputs {
-	#[default = false]
+	#[default(false)]
 	pub eww: bool,
 	/// Rate limit for eww updates. If set, eww updates will be batched and sent at most once per this duration.
 	pub eww_rate_limit: Option<Timeframe>,
-	#[default = true]
+	#[default(true)]
 	pub pipes: bool,
 	/// Per-line cap on how many not-yet-pushed values we remember. When a burst arrives faster than we
 	/// can flush, only the `buffer` most recent values are kept; older ones are dropped.
-	#[default = 16]
-	pub buffer: usize,
+	#[default(16)]
+	pub buffer: u8,
 	/// Backpressure: hard cap on the total number of queued-but-unflushed values across all lines. If a
 	/// new value would exceed this, we error out rather than let the queue grow unbounded.
-	#[default = 256]
-	pub max_flushes: usize,
+	#[default(64)]
+	pub max_flushes: u8,
 }
 
 #[derive(Clone, Debug, Default, MyConfigPrimitives, SettingsNested)]
